Collision, shit works.
the problem was __init__ needs two underlines. Run in ternimal to see
the printout.

I think im going to put the player moving stuff completely in the
player.py because that makes sense and would make updating position for
collision better.

The main file i think will just be a loader basically, it will load the
rockgroup1, doors1, and player for level one, and then rockgrounp2,
doors2, and player for level two etc.