Door works (almost)
Added rocks, they move when colliding with the door. The door steps
along an array that will have all the rooms. I will add Doors moving to
the same thing so hitting a door moves the play (as if entering from
the opposite side of the room) and a new door will be there