Enemy AI updated
The enemy AI now follows the player only if the distance between them is
less than 15. If it is greater, the enemy patrols in a square at the
position where distance became > 15.

Also added a helper.py file that currently only has a distance function.
We can use this file for any useful computations.